Move over from test_timedelta
