+++ python/cudf/cudf/tests/series/methods/test_reductions.py
@@ -11,8 +11,9 @@ import pytest
 from packaging.version import parse
 
 import cudf
+from cudf.core._compat import PANDAS_GE_230
 from cudf.testing import assert_eq
-from cudf.testing._utils import assert_exceptions_equal
+from cudf.testing._utils import assert_exceptions_equal, expect_warning_if
 
 
 @pytest.mark.parametrize("data", [[], [1, 2, 3]])
@@ -811,3 +812,100 @@ def test_categorical_reductions(request, reduction_methods):
     assert_exceptions_equal(
         getattr(psr, reduction_methods), getattr(gsr, reduction_methods)
     )
+
+
+@pytest.mark.parametrize(
+    "data_non_overflow",
+    [
+        [1000000, 200000, 3000000],
+        [1000000, 200000, None],
+        [],
+        [None],
+        [None, None, None, None, None],
+        [12, 12, 22, 343, 4353534, 435342],
+        np.array([10, 20, 30, None, 100]),
+        cp.asarray([10, 20, 30, 100]),
+        [1000000, 200000, 3000000],
+        [1000000, 200000, None],
+        [1],
+        [12, 11, 232, 223432411, 2343241, 234324, 23234],
+        [12, 11, 2.32, 2234.32411, 2343.241, 23432.4, 23234],
+        [1.321, 1132.324, 23223231.11, 233.41, 0.2434, 332, 323],
+        [12, 11, 2.32, 2234.32411, 2343.241, 23432.4, 23234],
+    ],
+)
+def test_timedelta_reduction_ops(
+    data_non_overflow, timedelta_types_as_str, reduction_methods
+):
+    if reduction_methods not in ["sum", "mean", "median", "quantile"]:
+        pytest.skip(f"{reduction_methods} not supported for timedelta")
+    gsr = cudf.Series(data_non_overflow, dtype=timedelta_types_as_str)
+    psr = gsr.to_pandas()
+
+    if len(psr) > 0 and psr.isnull().all() and reduction_methods == "median":
+        with pytest.warns(RuntimeWarning, match="Mean of empty slice"):
+            expected = getattr(psr, reduction_methods)()
+    else:
+        with expect_warning_if(
+            PANDAS_GE_230
+            and reduction_methods == "quantile"
+            and len(data_non_overflow) == 0
+            and timedelta_types_as_str != "timedelta64[ns]"
+        ):
+            expected = getattr(psr, reduction_methods)()
+    actual = getattr(gsr, reduction_methods)()
+    if pd.isna(expected) and pd.isna(actual):
+        pass
+    elif isinstance(expected, pd.Timedelta) and isinstance(
+        actual, pd.Timedelta
+    ):
+        assert (
+            expected.round(gsr._column.time_unit).value
+            == actual.round(gsr._column.time_unit).value
+        )
+    else:
+        assert_eq(expected, actual)
+
+
+@pytest.mark.parametrize("data", [[1, 2, 3], [], [1, 20, 1000, None]])
+@pytest.mark.parametrize("ddof", [1, 2, 3])
+def test_timedelta_std(data, timedelta_types_as_str, ddof):
+    gsr = cudf.Series(data, dtype=timedelta_types_as_str)
+    psr = gsr.to_pandas()
+
+    expected = psr.std(ddof=ddof)
+    actual = gsr.std(ddof=ddof)
+
+    if np.isnat(expected.to_numpy()) and np.isnat(actual.to_numpy()):
+        assert True
+    else:
+        np.testing.assert_allclose(
+            expected.to_numpy().astype("float64"),
+            actual.to_numpy().astype("float64"),
+            rtol=1e-5,
+            atol=0,
+        )
+
+
+@pytest.mark.parametrize("op", ["max", "min"])
+@pytest.mark.parametrize(
+    "data",
+    [
+        [],
+        [1, 2, 3, 100],
+        [10, None, 100, None, None],
+        [None, None, None],
+        [1231],
+    ],
+)
+def test_timedelta_reductions(data, op, timedelta_types_as_str):
+    sr = cudf.Series(data, dtype=timedelta_types_as_str)
+    psr = sr.to_pandas()
+
+    actual = getattr(sr, op)()
+    expected = getattr(psr, op)()
+
+    if np.isnat(expected.to_numpy()) and np.isnat(actual):
+        assert True
+    else:
+        assert_eq(expected.to_numpy(), actual)
